refactor(builder): move assembly Mode-specific parts into new mode submodule.
